moving display functions into one file
